pyproject.toml
[tool.poetry]
name = "rag-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pypdf = "^5.0.1"
langchain = "^0.3.3"
chromadb = "^0.5.15"
pytest = "^8.3.3"
boto3 = "^1.35.41"
langchain-community = "^0.3.2"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
